Oscillator params
